Move NearbyTransitFilterItemConstants to a different package.
